fix(agent): retarget the run reaper at the crates main actually names
The branch was written against `tinyagents::harness::`, a path that no
longer resolves: `vendor/tinyagents` is a virtual manifest now and the
member crates are named directly, so the sweep compiled only on its old
base. Point the three imports at `tinyagents_harness::`, matching how
`journal.rs` — which owns the status store this reaps through — imports
the same symbols today.

The layout gate also grew a rule since: inline `#[cfg(test)] mod tests`
is rejected in favour of a descriptive sibling. Move the sweep's tests
to `reaper_tests.rs` unchanged.
